backend/requirements.txt
<<<<<<< HEAD
altair==5.5.0
altgraph==0.17.5
anaconda-anon-usage @ file:///C:/b/abs_e8r_zga7xy/croot/anaconda-anon-usage_1732732454901/work
annotated-types @ file:///C:/b/abs_0dmaoyhhj3/croot/annotated-types_1709542968311/work
archspec @ file:///croot/archspec_1709217642129/work
asgiref==3.11.0
attrs==25.3.0
blinker==1.9.0
boltons @ file:///C:/b/abs_45_52ughkz/croot/boltons_1737061711836/work
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
cachetools==5.5.2
certifi @ file:///C:/b/abs_8a944p1_gn/croot/certifi_1738623753421/work/certifi
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1699472650914/work
comtypes==1.4.12
conda @ file:///C:/b/abs_12ybmrg1d4/croot/conda_1738168389704/work
conda-anaconda-telemetry @ file:///C:/b/abs_4c9llcc5ob/croot/conda-anaconda-telemetry_1736524617431/work
conda-anaconda-tos @ file:///C:/b/abs_ceeuq0lee_/croot/conda-anaconda-tos_1739299022910/work
conda-content-trust @ file:///C:/b/abs_bdfatn_wzf/croot/conda-content-trust_1714483201909/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1737733694612/work/src
conda-package-handling @ file:///C:/b/abs_7fz3aferfv/croot/conda-package-handling_1731369038903/work
conda_package_streaming @ file:///C:/b/abs_bdz9vbvbh2/croot/conda-package-streaming_1731366449946/work
contourpy==1.3.3
cryptography @ file:///C:/b/abs_e2lzchf4i6/croot/cryptography_1732130411942/work
cycler==0.12.1
distro @ file:///C:/b/abs_71xr36ua5r/croot/distro_1714488282676/work
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.15.1
filelock==3.18.0
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.59.2
frozendict @ file:///C:/b/abs_2alamqss6p/croot/frozendict_1713194885124/work
fsspec==2025.3.2
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gunicorn==21.2.0
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonpatch @ file:///C:/b/abs_4fdm88t7zi/croot/jsonpatch_1714483974578/work
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
libmambapy @ file:///C:/b/abs_627vsv8bhu/croot/mamba-split_1734469608328/work/libmambapy
markdown-it-py @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/markdown-it-py_1699473886965/work
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/mdurl_1699473506455/work
menuinst @ file:///C:/b/abs_fblttj5gp1/croot/menuinst_1738943438301/work
mpmath==1.3.0
narwhals==1.36.0
networkx==3.4.2
npm==0.1.1
npx==0.1.6
numpy==2.2.5
oauthlib==3.3.1
optional-django==0.1.0
packaging @ file:///C:/b/abs_3by6s2fa66/croot/packaging_1734472138782/work
pandas==2.2.3
pefile==2024.8.26
pillow==10.2.0
platformdirs @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/platformdirs_1701797392447/work
pluggy @ file:///C:/b/abs_dfec_m79vo/croot/pluggy_1733170145382/work
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycosat @ file:///C:/b/abs_18nblzzn70/croot/pycosat_1736868434419/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic @ file:///C:/b/abs_27dx58x550/croot/pydantic_1734736090499/work
pydantic_core @ file:///C:/b/abs_bdosz7qwys/croot/pydantic-core_1734726071532/work
pydeck==0.9.1
Pygments @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pygments_1699474141968/work
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
pyparsing==3.2.3
pypiwin32==223
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
python-dateutil==2.9.0.post0
pyttsx3==2.99
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
referencing==0.36.2
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
requests-oauthlib==2.0.0
rich @ file:///C:/b/abs_8b92fzmygg/croot/rich_1732638991774/work
rpds-py==0.24.0
rsa==4.9.1
ruamel.yaml @ file:///C:/b/abs_0cunwx_ww6/croot/ruamel.yaml_1727980181547/work
ruamel.yaml.clib @ file:///C:/b/abs_5fk8zi6n09/croot/ruamel.yaml.clib_1727769837359/work
scikit-learn==1.7.1
scipy==1.16.1
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sqlparse==0.5.4
streamlit==1.44.1
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm @ file:///C:/b/abs_0eh9b6xugj/croot/tqdm_1738945553987/work
truststore @ file:///C:/b/abs_494cm143zh/croot/truststore_1736550137835/work
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata==2025.2
urllib3 @ file:///C:/b/abs_7bst06lizn/croot/urllib3_1737133657081/work
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.6.0
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
zstandard @ file:///C:/b/abs_31t8xmrv_h/croot/zstandard_1731356578015/work
=======
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.4
whitenoise==6.11.0
>>>>>>> fa68af2688e944f226ee83eb76320dfc73dcd9f1
